trim description
